fixed cmake macros
git-svn-id: https://camplinux.in.tum.de/svn/campvis/trunk@178 bb408c1c-ae56-11e1-83d9-df6b3e0c105e

+++ src/tumvis/core/datastructures/imagedatadisk.cpp
@@ -39,7 +39,7 @@ namespace TUMVis {
         // open file and prepare for read
         std::ifstream file(_url.c_str(), std::ios::in | std::ios::binary | std::ios::ate);
         if (file.is_open()) {
-            size_t fileSize = file.tellg();
+            size_t fileSize = static_cast<size_t>(file.tellg());
             if (fileSize < numBytes) {
                 LERROR("File is smaller than expected.");
                 return WeaklyTypedPointer(_type, 0);
